my-dummy-pipeline/server_predictor/requirements.txt
scikit-learn
joblib
google-cloud-storage
#vp_abstractor